requirements.txt
cryptography==3.2.1
numpy==1.20.1
pandas==1.2.2
pycparser==2.20
pycryptodome==3.11.0
pylint==2.6.0
scikit-learn==1.0.1
scipy==1.7.2
tqdm==4.62.3
virtualenv==20.0.31
virtualenvwrapper-win==1.2.6

